chore: rename pox-delegation to pox-pools-1-cycle
