+++ Clarinet.toml
@@ -5,8 +5,8 @@ authors = []
 telemetry = true
 cache_dir = './.cache'
 
-[contracts.pox-delegation]
-path = 'contracts/pox-delegation.clar'
+[contracts.pox-pools-1-cycle]
+path = 'contracts/pox-pools-1-cycle.clar'
 clarity_version = 2
 epoch = 2.1
