+++ coverage-web/contracts/index-sort-f.html
@@ -31,18 +31,18 @@
             <td class="headerValue">coverage.lcov</td>
             <td></td>
             <td class="headerItem">Lines:</td>
-            <td class="headerCovTableEntry">68</td>
-            <td class="headerCovTableEntry">68</td>
-            <td class="headerCovTableEntryHi">100.0 %</td>
+            <td class="headerCovTableEntry">150</td>
+            <td class="headerCovTableEntry">228</td>
+            <td class="headerCovTableEntryLo">65.8 %</td>
           </tr>
           <tr>
             <td class="headerItem">Date:</td>
-            <td class="headerValue">2023-03-09 18:56:44</td>
+            <td class="headerValue">2023-03-20 11:14:42</td>
             <td></td>
             <td class="headerItem">Functions:</td>
-            <td class="headerCovTableEntry">17</td>
-            <td class="headerCovTableEntry">17</td>
-            <td class="headerCovTableEntryHi">100.0 %</td>
+            <td class="headerCovTableEntry">34</td>
+            <td class="headerCovTableEntry">53</td>
+            <td class="headerCovTableEntryLo">64.2 %</td>
           </tr>
           <tr><td><img src="../glass.png" width=3 height=3 alt=""></td></tr>
         </table>
@@ -70,14 +70,24 @@
       <td class="tableHead" colspan=2>Functions <span class="tableHeadSort"><img src="../glass.png" width=10 height=14 alt="Sort by function coverage" title="Sort by function coverage" border=0></span></td>
     </tr>
     <tr>
-      <td class="coverFile"><a href="pox-delegation.clar.gcov.html">pox-delegation.clar</a></td>
+      <td class="coverFile"><a href="fp-delegation.clar.gcov.html">fp-delegation.clar</a></td>
       <td class="coverBar" align="center">
-        <table border=0 cellspacing=0 cellpadding=1><tr><td class="coverBarOutline"><img src="../emerald.png" width=100 height=10 alt="100.0%"></td></tr></table>
+        <table border=0 cellspacing=0 cellpadding=1><tr><td class="coverBarOutline"><img src="../ruby.png" width=57 height=10 alt="57.3%"><img src="../snow.png" width=43 height=10 alt="57.3%"></td></tr></table>
       </td>
-      <td class="coverPerHi">100.0&nbsp;%</td>
-      <td class="coverNumHi">68 / 68</td>
-      <td class="coverPerHi">100.0&nbsp;%</td>
-      <td class="coverNumHi">17 / 17</td>
+      <td class="coverPerLo">57.3&nbsp;%</td>
+      <td class="coverNumLo">67 / 117</td>
+      <td class="coverPerLo">44.8&nbsp;%</td>
+      <td class="coverNumLo">13 / 29</td>
+    </tr>
+    <tr>
+      <td class="coverFile"><a href="pox-pools-1-cycle.clar.gcov.html">pox-pools-1-cycle.clar</a></td>
+      <td class="coverBar" align="center">
+        <table border=0 cellspacing=0 cellpadding=1><tr><td class="coverBarOutline"><img src="../ruby.png" width=75 height=10 alt="74.8%"><img src="../snow.png" width=25 height=10 alt="74.8%"></td></tr></table>
+      </td>
+      <td class="coverPerLo">74.8&nbsp;%</td>
+      <td class="coverNumLo">83 / 111</td>
+      <td class="coverPerMed">87.5&nbsp;%</td>
+      <td class="coverNumMed">21 / 24</td>
     </tr>
   </table>
   </center>
