--- coverage-web/contracts/pox-delegation.clar.func-sort-c.html
@@ -1,140 +1,0 @@
-<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN">
-
-<html lang="en">
-
-<head>
-  <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-  <title>LCOV - coverage.lcov - contracts/pox-delegation.clar - functions</title>
-  <link rel="stylesheet" type="text/css" href="../gcov.css">
-</head>
-
-<body>
-
-  <table width="100%" border=0 cellspacing=0 cellpadding=0>
-    <tr><td class="title">LCOV - code coverage report</td></tr>
-    <tr><td class="ruler"><img src="../glass.png" width=3 height=3 alt=""></td></tr>
-
-    <tr>
-      <td width="100%">
-        <table cellpadding=1 border=0 width="100%">
-          <tr>
-            <td width="10%" class="headerItem">Current view:</td>
-            <td width="35%" class="headerValue"><a href="../index.html">top level</a> - <a href="index.html">contracts</a> - pox-delegation.clar<span style="font-size: 80%;"> (<a href="pox-delegation.clar.gcov.html">source</a> / functions)</span></td>
-            <td width="5%"></td>
-            <td width="15%"></td>
-            <td width="10%" class="headerCovTableHead">Hit</td>
-            <td width="10%" class="headerCovTableHead">Total</td>
-            <td width="15%" class="headerCovTableHead">Coverage</td>
-          </tr>
-          <tr>
-            <td class="headerItem">Test:</td>
-            <td class="headerValue">coverage.lcov</td>
-            <td></td>
-            <td class="headerItem">Lines:</td>
-            <td class="headerCovTableEntry">68</td>
-            <td class="headerCovTableEntry">68</td>
-            <td class="headerCovTableEntryHi">100.0 %</td>
-          </tr>
-          <tr>
-            <td class="headerItem">Date:</td>
-            <td class="headerValue">2023-03-09 18:56:44</td>
-            <td></td>
-            <td class="headerItem">Functions:</td>
-            <td class="headerCovTableEntry">17</td>
-            <td class="headerCovTableEntry">17</td>
-            <td class="headerCovTableEntryHi">100.0 %</td>
-          </tr>
-          <tr><td><img src="../glass.png" width=3 height=3 alt=""></td></tr>
-        </table>
-      </td>
-    </tr>
-
-    <tr><td class="ruler"><img src="../glass.png" width=3 height=3 alt=""></td></tr>
-  </table>
-
-  <center>
-  <table width="60%" cellpadding=1 cellspacing=1 border=0>
-    <tr><td><br></td></tr>
-    <tr>
-      <td width="80%" class="tableHead">Function Name <span class="tableHeadSort"><a href="pox-delegation.clar.func.html"><img src="../updown.png" width=10 height=14 alt="Sort by function name" title="Sort by function name" border=0></a></span></td>
-      <td width="20%" class="tableHead">Hit count <span class="tableHeadSort"><img src="../glass.png" width=10 height=14 alt="Sort by hit count" title="Sort by hit count" border=0></span></td>
-    </tr>
-    <tr>
-              <td class="coverFn"><a href="pox-delegation.clar.gcov.html#62">asserts-panic</a></td>
-              <td class="coverFnHi">2</td>
-            </tr>
-    <tr>
-              <td class="coverFn"><a href="pox-delegation.clar.gcov.html#77">insert-in-new-list</a></td>
-              <td class="coverFnHi">2</td>
-            </tr>
-    <tr>
-              <td class="coverFn"><a href="pox-delegation.clar.gcov.html#176">get-user-data</a></td>
-              <td class="coverFnHi">5</td>
-            </tr>
-    <tr>
-              <td class="coverFn"><a href="pox-delegation.clar.gcov.html#171">get-status-list</a></td>
-              <td class="coverFnHi">6</td>
-            </tr>
-    <tr>
-              <td class="coverFn"><a href="pox-delegation.clar.gcov.html#158">get-status</a></td>
-              <td class="coverFnHi">9</td>
-            </tr>
-    <tr>
-              <td class="coverFn"><a href="pox-delegation.clar.gcov.html#43">pox-get-stacker-info</a></td>
-              <td class="coverFnHi">9</td>
-            </tr>
-    <tr>
-              <td class="coverFn"><a href="pox-delegation.clar.gcov.html#82">map-set-details</a></td>
-              <td class="coverFnHi">12</td>
-            </tr>
-    <tr>
-              <td class="coverFn"><a href="pox-delegation.clar.gcov.html#69">merge-details</a></td>
-              <td class="coverFnHi">12</td>
-            </tr>
-    <tr>
-              <td class="coverFn"><a href="pox-delegation.clar.gcov.html#141">delegate-stack-stx</a></td>
-              <td class="coverFnHi">14</td>
-            </tr>
-    <tr>
-              <td class="coverFn"><a href="pox-delegation.clar.gcov.html#57">min</a></td>
-              <td class="coverFnHi">14</td>
-            </tr>
-    <tr>
-              <td class="coverFn"><a href="pox-delegation.clar.gcov.html#95">pox-delegate-stack-stx</a></td>
-              <td class="coverFnHi">14</td>
-            </tr>
-    <tr>
-              <td class="coverFn"><a href="pox-delegation.clar.gcov.html#165">get-status-lists-last-index</a></td>
-              <td class="coverFnHi">16</td>
-            </tr>
-    <tr>
-              <td class="coverFn"><a href="pox-delegation.clar.gcov.html#182">get-total</a></td>
-              <td class="coverFnHi">26</td>
-            </tr>
-    <tr>
-              <td class="coverFn"><a href="pox-delegation.clar.gcov.html#131">delegate-stx</a></td>
-              <td class="coverFnHi">84</td>
-            </tr>
-    <tr>
-              <td class="coverFn"><a href="pox-delegation.clar.gcov.html#47">pox-delegate-stx</a></td>
-              <td class="coverFnHi">84</td>
-            </tr>
-    <tr>
-              <td class="coverFn"><a href="pox-delegation.clar.gcov.html#30">burn-height-to-reward-cycle</a></td>
-              <td class="coverFnHi">96</td>
-            </tr>
-    <tr>
-              <td class="coverFn"><a href="pox-delegation.clar.gcov.html#38">current-pox-reward-cycle</a></td>
-              <td class="coverFnHi">96</td>
-            </tr>
-  </table>
-  <br>
-  </center>
-  <table width="100%" border=0 cellspacing=0 cellpadding=0>
-    <tr><td class="ruler"><img src="../glass.png" width=3 height=3 alt=""></td></tr>
-    <tr><td class="versionInfo">Generated by: <a href="http://ltp.sourceforge.net/coverage/lcov.php" target="_parent">LCOV version 1.14</a></td></tr>
-  </table>
-  <br>
-
-</body>
-</html>
