+++ coverage-web/contracts/testing/helper.clar.gcov.html
@@ -1,0 +1,93 @@
+<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN">
+
+<html lang="en">
+
+<head>
+  <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
+  <title>LCOV - coverage.lcov - contracts/testing/helper.clar</title>
+  <link rel="stylesheet" type="text/css" href="../../gcov.css">
+</head>
+
+<body>
+
+  <table width="100%" border=0 cellspacing=0 cellpadding=0>
+    <tr><td class="title">LCOV - code coverage report</td></tr>
+    <tr><td class="ruler"><img src="../../glass.png" width=3 height=3 alt=""></td></tr>
+
+    <tr>
+      <td width="100%">
+        <table cellpadding=1 border=0 width="100%">
+          <tr>
+            <td width="10%" class="headerItem">Current view:</td>
+            <td width="35%" class="headerValue"><a href="../../index.html">top level</a> - <a href="index.html">contracts/testing</a> - helper.clar<span style="font-size: 80%;"> (source / <a href="helper.clar.func-sort-c.html">functions</a>)</span></td>
+            <td width="5%"></td>
+            <td width="15%"></td>
+            <td width="10%" class="headerCovTableHead">Hit</td>
+            <td width="10%" class="headerCovTableHead">Total</td>
+            <td width="15%" class="headerCovTableHead">Coverage</td>
+          </tr>
+          <tr>
+            <td class="headerItem">Test:</td>
+            <td class="headerValue">coverage.lcov</td>
+            <td></td>
+            <td class="headerItem">Lines:</td>
+            <td class="headerCovTableEntry">1</td>
+            <td class="headerCovTableEntry">4</td>
+            <td class="headerCovTableEntryLo">25.0 %</td>
+          </tr>
+          <tr>
+            <td class="headerItem">Date:</td>
+            <td class="headerValue">2023-03-20 11:14:42</td>
+            <td></td>
+            <td class="headerItem">Functions:</td>
+            <td class="headerCovTableEntry">1</td>
+            <td class="headerCovTableEntry">4</td>
+            <td class="headerCovTableEntryLo">25.0 %</td>
+          </tr>
+          <tr><td><img src="../../glass.png" width=3 height=3 alt=""></td></tr>
+        </table>
+      </td>
+    </tr>
+
+    <tr><td class="ruler"><img src="../../glass.png" width=3 height=3 alt=""></td></tr>
+  </table>
+
+  <table cellpadding=0 cellspacing=0 border=0>
+    <tr>
+      <td><br></td>
+    </tr>
+    <tr>
+      <td>
+<pre class="sourceHeading">          Line data    Source code</pre>
+<pre class="source">
+<a name="1"><span class="lineNum">       1 </span>            : ;; Helper contract mainly to wrap read-only function in public functions</a>
+<a name="2"><span class="lineNum">       2 </span>            : ;; that can be used with &quot;clarinet integrate&quot;.</a>
+<a name="3"><span class="lineNum">       3 </span>            : ;; Currently, it is not possible to use read-only functions in deployment plans.</a>
+<a name="4"><span class="lineNum">       4 </span>            : </a>
+<a name="5"><span class="lineNum">       5 </span>            : (define-public (get-user-data (user principal))</a>
+<a name="6"><span class="lineNum">       6 </span><span class="lineNoCov">          0 :   (ok (print (contract-call? .pox-pools-1-cycle get-user-data user))))</span></a>
+<a name="7"><span class="lineNum">       7 </span>            : </a>
+<a name="8"><span class="lineNum">       8 </span>            : (define-public (get-stx-account (user principal))</a>
+<a name="9"><span class="lineNum">       9 </span><span class="lineNoCov">          0 :   (ok (print (contract-call? .pox-pools-1-cycle get-stx-account user))))</span></a>
+<a name="10"><span class="lineNum">      10 </span>            : </a>
+<a name="11"><span class="lineNum">      11 </span>            : (define-public (get-reward-set (reward-cycle uint))</a>
+<a name="12"><span class="lineNum">      12 </span><span class="lineNoCov">          0 :   (ok (print (contract-call? .fp-delegation get-reward-set reward-cycle))))</span></a>
+<a name="13"><span class="lineNum">      13 </span>            : </a>
+<a name="14"><span class="lineNum">      14 </span>            : ;; call wrapper contract from a contract</a>
+<a name="15"><span class="lineNum">      15 </span>            : ;; requires allowance</a>
+<a name="16"><span class="lineNum">      16 </span>            : (define-public (delegate-stx (amount-ustx uint))</a>
+<a name="17"><span class="lineNum">      17 </span><span class="lineCov">          6 :   (contract-call? .fp-delegation delegate-stx amount-ustx))</span></a>
+</pre>
+      </td>
+    </tr>
+  </table>
+  <br>
+
+  <table width="100%" border=0 cellspacing=0 cellpadding=0>
+    <tr><td class="ruler"><img src="../../glass.png" width=3 height=3 alt=""></td></tr>
+    <tr><td class="versionInfo">Generated by: <a href="http://ltp.sourceforge.net/coverage/lcov.php" target="_parent">LCOV version 1.14</a></td></tr>
+  </table>
+  <br>
+
+</body>
+</html>
