+++ coverage-web/index.html
@@ -31,18 +31,18 @@
             <td class="headerValue">coverage.lcov</td>
             <td></td>
             <td class="headerItem">Lines:</td>
-            <td class="headerCovTableEntry">68</td>
-            <td class="headerCovTableEntry">68</td>
-            <td class="headerCovTableEntryHi">100.0 %</td>
+            <td class="headerCovTableEntry">151</td>
+            <td class="headerCovTableEntry">232</td>
+            <td class="headerCovTableEntryLo">65.1 %</td>
           </tr>
           <tr>
             <td class="headerItem">Date:</td>
-            <td class="headerValue">2023-03-09 18:56:44</td>
+            <td class="headerValue">2023-03-20 11:14:42</td>
             <td></td>
             <td class="headerItem">Functions:</td>
-            <td class="headerCovTableEntry">17</td>
-            <td class="headerCovTableEntry">17</td>
-            <td class="headerCovTableEntryHi">100.0 %</td>
+            <td class="headerCovTableEntry">35</td>
+            <td class="headerCovTableEntry">57</td>
+            <td class="headerCovTableEntryLo">61.4 %</td>
           </tr>
           <tr><td><img src="glass.png" width=3 height=3 alt=""></td></tr>
         </table>
@@ -72,12 +72,22 @@
     <tr>
       <td class="coverFile"><a href="contracts/index.html">contracts</a></td>
       <td class="coverBar" align="center">
-        <table border=0 cellspacing=0 cellpadding=1><tr><td class="coverBarOutline"><img src="emerald.png" width=100 height=10 alt="100.0%"></td></tr></table>
+        <table border=0 cellspacing=0 cellpadding=1><tr><td class="coverBarOutline"><img src="ruby.png" width=66 height=10 alt="65.8%"><img src="snow.png" width=34 height=10 alt="65.8%"></td></tr></table>
       </td>
-      <td class="coverPerHi">100.0&nbsp;%</td>
-      <td class="coverNumHi">68 / 68</td>
-      <td class="coverPerHi">100.0&nbsp;%</td>
-      <td class="coverNumHi">17 / 17</td>
+      <td class="coverPerLo">65.8&nbsp;%</td>
+      <td class="coverNumLo">150 / 228</td>
+      <td class="coverPerLo">64.2&nbsp;%</td>
+      <td class="coverNumLo">34 / 53</td>
+    </tr>
+    <tr>
+      <td class="coverFile"><a href="contracts/testing/index.html">contracts/testing</a></td>
+      <td class="coverBar" align="center">
+        <table border=0 cellspacing=0 cellpadding=1><tr><td class="coverBarOutline"><img src="ruby.png" width=25 height=10 alt="25.0%"><img src="snow.png" width=75 height=10 alt="25.0%"></td></tr></table>
+      </td>
+      <td class="coverPerLo">25.0&nbsp;%</td>
+      <td class="coverNumLo">1 / 4</td>
+      <td class="coverPerLo">25.0&nbsp;%</td>
+      <td class="coverNumLo">1 / 4</td>
     </tr>
   </table>
   </center>
