+++ tests/pox-pools-1-cycle_flow_test.ts
@@ -7,7 +7,7 @@ import {
   delegateStackStx,
   delegateStx,
   getStatus,
-} from "./client/pox-delegation-client.ts";
+} from "./client/pox-pools-1-cycle-client.ts";
 import { Clarinet, Tx, Chain, Account, types } from "./deps.ts";
 import {
   btcAddrWallet1,
@@ -23,7 +23,7 @@ Clarinet.test({
     let deployer = accounts.get("deployer")!;
     let wallet_1 = accounts.get("wallet_1")!;
     let wallet_2 = accounts.get("wallet_2")!;
-    const poxDelegationContract = deployer.address + ".pox-delegation";
+    const poxDelegationContract = deployer.address + ".pox-pools-1-cycle";
 
     let block = chain.mineBlock([
       allowContractCaller(poxDelegationContract, undefined, deployer),
@@ -101,7 +101,7 @@ Clarinet.test({
     let wallet_1 = accounts.get("wallet_1")!;
     let wallet_2 = accounts.get("wallet_2")!;
     let faucet = accounts.get("faucet")!;
-    const poxDelegationContract = deployer.address + ".pox-delegation";
+    const poxDelegationContract = deployer.address + ".pox-pools-1-cycle";
 
     let block = chain.mineBlock([
       allowContractCaller(poxDelegationContract, undefined, deployer),
@@ -198,7 +198,7 @@ Clarinet.test({
     let deployer = accounts.get("deployer")!;
     let wallet_1 = accounts.get("wallet_1")!;
     let wallet_2 = accounts.get("wallet_2")!;
-    const poxDelegationContract = deployer.address + ".pox-delegation";
+    const poxDelegationContract = deployer.address + ".pox-pools-1-cycle";
 
     let block = chain.mineBlock([
       allowContractCaller(poxDelegationContract, undefined, deployer),
@@ -269,7 +269,7 @@ Clarinet.test({
   async fn(chain: Chain, accounts: Map<string, Account>) {
     let deployer = accounts.get("deployer")!;
     let wallet_1 = accounts.get("wallet_1")!;
-    const poxDelegationContract = deployer.address + ".pox-delegation";
+    const poxDelegationContract = deployer.address + ".pox-pools-1-cycle";
 
     let block = chain.mineBlock([
       allowContractCaller(poxDelegationContract, undefined, deployer),
@@ -343,7 +343,7 @@ Clarinet.test({
     console.log(
       "*** stx-account",
       chain.callReadOnlyFn(
-        "pox-delegation",
+        "pox-pools-1-cycle",
         "get-stx-account",
         [types.principal(wallet_1.address)],
         wallet_1.address
