+++ tests/pox-pools-1-cycle_test.ts
@@ -7,7 +7,7 @@ import {
   delegateStx,
   getStatus,
   getUserData,
-} from "./client/pox-delegation-client.ts";
+} from "./client/pox-pools-1-cycle-client.ts";
 import { Clarinet, Tx, Chain, Account, types } from "./deps.ts";
 import {
   Errors,
@@ -22,7 +22,7 @@ Clarinet.test({
   async fn(chain: Chain, accounts: Map<string, Account>) {
     const deployer = accounts.get("deployer")!;
     const wallet_1 = accounts.get("wallet_1")!;
-    const poxDelegationContract = deployer.address + ".pox-delegation";
+    const poxDelegationContract = deployer.address + ".pox-pools-1-cycle";
 
     let block = chain.mineBlock([
       allowContractCaller(poxDelegationContract, undefined, wallet_1),
@@ -70,7 +70,7 @@ Clarinet.test({
     let deployer = accounts.get("deployer")!;
     let wallet_1 = accounts.get("wallet_1")!;
     let wallet_2 = accounts.get("wallet_2")!;
-    const poxDelegationContract = deployer.address + ".pox-delegation";
+    const poxDelegationContract = deployer.address + ".pox-pools-1-cycle";
 
     let block = chain.mineBlock([
       allowContractCaller(poxDelegationContract, undefined, wallet_1),
@@ -118,7 +118,7 @@ Clarinet.test({
   async fn(chain: Chain, accounts: Map<string, Account>) {
     let deployer = accounts.get("deployer")!;
     let wallet_1 = accounts.get("wallet_1")!;
-    const poxDelegationContract = deployer.address + ".pox-delegation";
+    const poxDelegationContract = deployer.address + ".pox-pools-1-cycle";
 
     let block = chain.mineBlock([
       allowContractCaller(poxDelegationContract, undefined, deployer),
